Subtitle entry switching
On line 28, add a new variable called subEntry.  This allows the ability to switch between Sub entry 2, 3, and 4 with ease.  All the Subtitle functions on line 270 have commands that check what value subEntry has, and adds each subtitle accordingly.  This currently does not have any effect on VBoy subtitles as of now.